refactor(runners): centralize provider contract
