+++ lambdas/functions/control-plane/src/pool/pool-provider.ts
@@ -1,6 +1,6 @@
 import type { Octokit } from '@octokit/rest';
+import type { RunnerProvider } from '@aws-github-runner/runner-provider';
 
-import type { RunnerProvider } from '../runner-provider';
 import type { CreateGitHubRunnerConfig, GitHubRunnerType } from '../scale-runners/types';
 
 export interface RunnerStatus {
