+++ lambdas/functions/control-plane/src/scale-runners/scale-down.ts
@@ -2,11 +2,11 @@ import { Octokit } from '@octokit/rest';
 import { Endpoints } from '@octokit/types';
 import { RequestError } from '@octokit/request-error';
 import { createChildLogger } from '@aws-github-runner/aws-powertools-util';
+import { resolveRunnerProviderType } from '@aws-github-runner/runner-provider';
 import moment from 'moment';
 
 import { createGithubAppAuth, createGithubInstallationAuth, createOctokitClient } from '../github/auth';
 import { createScaleDownRunnerProvider } from '../runner-provider-registry';
-import { resolveRunnerProviderType } from '../runner-provider';
 import { GhRunners, githubCache } from './cache';
 import { ScalingDownConfigList, getEvictionStrategy, getIdleRunnerCount } from './scale-down-config';
 import { metricGitHubAppRateLimit } from '../github/rate-limit';
